third_party/WebKit/LayoutTests/bluetooth/requestDevice/acceptAllDevices/device-with-uuids.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('NoNameHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({acceptAllDevices: true}))
    .then(device => {
      assert_true(device.name === null);
    });
}, 'Device with only UUIDs nearby. Should be found if ' +
   'acceptAllDevices is true.');
</script>
